+++ controller/Changelog.md
@@ -18,6 +18,10 @@
 
 - os: Remove unnecessary administration capabilities for hardening
 
+## Changed
+
+- os: Ignore suspend and hibernation key presses, but interpret as poweroff when long-pressed
+
 # [2023.9.1] - 2024-03-15
 
 # [2023.9.1-VALIDATION] - 2024-03-12